Conflicts:
	bank/bank-client/src/app/app.module.ts
	bank/bank-client/src/app/components/home/home.component.html
	bank/bank-client/src/app/components/home/home.component.ts
	bank/src/main/java/com/example/bank/repository/BankAccountRepository.java
	bank/src/main/java/com/example/bank/service/BankAccountService.java
	bank/src/main/java/com/example/bank/service/PlaceService.java
	bank/src/main/java/com/example/bank/service/impl/PlaceServiceImpl.java

+++ bank/src/main/java/com/example/bank/controller/BankAccountController.java
@@ -1,25 +1,50 @@
 package com.example.bank.controller;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.OutputStream;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.text.SimpleDateFormat;  
 import java.util.Date; 
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+
+import com.example.bank.model.AccountForBank;
+import com.example.bank.model.AnalyticsOfStatements;
+import com.example.bank.model.Bank;
 import com.example.bank.DTO.BankAccountDTO;
 import com.example.bank.model.BankAccount;
+import com.example.bank.model.DailyAccountBalance;
+import com.example.bank.model.PaymentType;
+import com.example.bank.model.Place;
+import com.example.bank.service.AnalyticsOfStatementsService;
 import com.example.bank.service.BankAccountService;
 import com.example.bank.service.BankService;
+import com.example.bank.service.DailyAccountBalanceService;
+import com.example.bank.service.PaymentTypeService;
+import com.example.bank.service.PlaceService;
 import com.example.bank.service.ClientService;
 import com.example.bank.service.CurrencyService;
 
+//import net.sf.jasperreports.engine.JREmptyDataSource;
+//import net.sf.jasperreports.engine.JasperExportManager;
+//import net.sf.jasperreports.engine.JasperFillManager;
+//import net.sf.jasperreports.engine.JasperPrint;
+//import net.sf.jasperreports.engine.data.JRBeanCollectionDataSource;
+// 
 @RestController
 @CrossOrigin(origins = "*")
 @RequestMapping("/public/bankAccounts")
@@ -32,10 +57,21 @@ public class BankAccountController {
 	private BankService bankService;
 	
 	@Autowired
+	private DailyAccountBalanceService dailyAccountBalanceService;
+
 	private ClientService clientService;
 	
 	@Autowired
 	private CurrencyService currencyService;
+	
+	@Autowired
+	private AnalyticsOfStatementsService analyticService;
+	
+	@Autowired
+	private PaymentTypeService paymentTypeService;
+	
+	@Autowired
+	private PlaceService placeService;
 	
 	
 	@RequestMapping(
@@ -92,16 +128,92 @@ public class BankAccountController {
 	@RequestMapping(
 			value = "suspendAccount/{id}/{trasferAccount}",
 			method = RequestMethod.DELETE)			
-	public List<BankAccount> suspendAccount(@PathVariable("id") Long id, @PathVariable("trasferAccount") Long transferAccount){
+	public List<BankAccount> suspendAccount(@PathVariable("id") Long id, @PathVariable("trasferAccount") String transferAccount){
+		
+		System.out.println("TRANSFER NUMBER JE ::::: " + transferAccount);
+		
 		double trasferedMoney = bankAccountService.findById(id).getMoney();
-		BankAccount trasfer = bankAccountService.findById(transferAccount);
+		BankAccount trasfer = bankAccountService.findByAccNumber(transferAccount);
 		
-		trasfer.setMoney(trasfer.getMoney() + trasferedMoney);
-		bankAccountService.save(trasfer);
 		
-		bankAccountService.deleteById(id);
+		BankAccount origin = bankAccountService.findById(id);
 		
+		long time = System.currentTimeMillis();
+		java.sql.Date currentDate = new java.sql.Date(time);
+		
+		boolean emergencyBool = false;
+		if(trasfer.getMoney() > 250000.0) emergencyBool = true;
+		
+		DailyAccountBalance dailyAccountBalance = dailyAccountBalanceService.findByAccountNumberAndDate(origin, currentDate);
+		PaymentType pt = paymentTypeService.findById(2l);
+		
+		
+		if(trasfer.getBank().getName().equals(origin.getBank().getName())) {
+			
+			trasfer.setMoney(trasfer.getMoney() + trasferedMoney);
+			bankAccountService.save(trasfer);
+			
+			bankAccountService.deleteById(id);
+			
+			
+		}
+		else {
+			
+			
+			//	AnalyticsOfStatements nova = new AnalyticsOfStatements(origin.getClient().getName(), "Tansfer", trasfer.getClient().getName(), currentDate, currentDate, origin.getAccountNumber(), null, 
+			//		"987-446-587", trasfer.getAccountNumber(), null, "4654-6216", emergencyBool, trasfer.getMoney(), null, null, dailyAccountBalance,null, null, null, null);
+		
+//			AnalyticsOfStatements nova = new AnalyticsOfStatements(origin.getClient().getName(), "Tansfer", trasfer.getClient().getName(), currentDate, currentDate, origin.getAccountNumber(), 97,
+//					"987-446-587", trasfer.getAccountNumber(), 15, "4654-6216", emergencyBool, (float)trasfer.getMoney(), 1, "0", dailyAccountBalance, pt, currencyService.getCurrencyById(1l), null, placeService.findById(1l));
+//			analyticService.save(nova);
+//			bankAccountService.deleteById(id);
+		}
+		
+		//97 15 1 0 2 1
 		return bankAccountService.getAll();
+	}
+	
+	// NOVO
+	
+	@GetMapping("/pdf/{id}")
+	public void exportAccountsToPdf(@PathVariable Long id) throws Exception {
+		ArrayList<AccountForBank> list = new ArrayList<>();
+
+		ArrayList<BankAccount> accounts = bankAccountService.findByBank(id);
+		Bank bank = bankService.findOne(id);
+		Date date = new Date();
+		for (BankAccount account : accounts) {
+			String clientName = "";
+			if (account.getClient().getTypeOfClient().equals("fizicko lice")) {
+				clientName = account.getClient().getName();
+			} else {
+				//clientName = ((BankAccount) account.getClient()).getNaziv_klijenta();
+				clientName = account.getClient().getName();
+			}
+			DailyAccountBalance dailyAccountBalance = dailyAccountBalanceService.findByAccountNumberAndDate(account, new Date());
+			Float state = (float) 0.0;
+			if (dailyAccountBalance.getTrafficToBenefit() == 0 && dailyAccountBalance.getTrafficToTheBurden() == 0)
+				state = dailyAccountBalance.getPreviousState();
+			else
+				state = dailyAccountBalance.getNewState();
+			AccountForBank afb = new AccountForBank(clientName, account.getAccountNumber(), state.toString(),
+					account.getCurrency().getOfficial_code());
+			list.add(afb);
+		}
+
+	/*	JRBeanCollectionDataSource itemsJRBean = new JRBeanCollectionDataSource(list);
+
+
+		Map<String, Object> parameters = new HashMap<String, Object>();
+		parameters.put("ItemDataSource", itemsJRBean);
+		parameters.put("BankName", bank.getName());
+		parameters.put("Date", date);
+		
+		JasperPrint jasperPrint = JasperFillManager.fillReport("excerptBank.jasper", parameters, new JREmptyDataSource());
+	    File file = new File("../bank/accounts.pdf");
+	    OutputStream outputStream = new FileOutputStream(file);
+	    JasperExportManager.exportReportToPdfStream(jasperPrint, outputStream); */
+	
 	}
 
 }
